# libs/api-client-go/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@daytona.com
    name: Daytona Platforms Inc.
    url: https://www.daytona.io
  description: Daytona AI platform API Docs
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  title: Daytona
  version: "1.0"
servers:
- url: http://localhost:3000
paths:
...
  /sandbox/{sandboxIdOrName}/build-logs-url:
    get:
      operationId: getBuildLogsUrl
      parameters:
      - explode: false
        in: path
        name: sandboxIdOrName
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Url"
          description: ""
      summary: Get build logs URL
      tags:
      - sandbox
...
  /snapshots:
    get:
      operationId: getAllSnapshots
      parameters:
      - description: Page number of the results
        explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          type: number
        style: form
      - description: Number of results per page
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 100
          maximum: 200
          minimum: 1
          type: number
        style: form
      - description: Filter by partial name match
        explode: true
        in: query
        name: name
        required: false
        schema:
          example: abc123
          type: string
        style: form
      - description: Field to sort by
        explode: true
        in: query
        name: sort
        required: false
        schema:
          default: lastUsedAt
          enum:
          - name
          - state
          - lastUsedAt
          - createdAt
          type: string
        style: form
      - description: Direction to sort by
        explode: true
        in: query
        name: order
        required: false
        schema:
          default: desc
          enum:
          - asc
          - desc
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedSnapshots"
          description: Paginated list of all snapshots
      summary: List all snapshots
      tags:
      - snapshots
    post:
      operationId: createSnapshot
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSnapshot"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SnapshotDto"
          description: The snapshot has been successfully created.
        "400":
          description: Bad request - Snapshots with tag ":latest" are not allowed
      summary: Create a new snapshot
      tags:
      - snapshots
  /snapshots/{id}:
    delete:
      operationId: removeSnapshot
      parameters:
      - description: Snapshot ID
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Snapshot has been deleted
      summary: Delete snapshot
      tags:
      - snapshots
    get:
      operationId: getSnapshot
      parameters:
      - description: Snapshot ID or name
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SnapshotDto"
          description: The snapshot
        "404":
          description: Snapshot not found
      summary: Get snapshot by ID or name
      tags:
      - snapshots
  /snapshots/{id}/build-logs:
    get:
      deprecated: true
      description: This endpoint is deprecated. Use `getSnapshotBuildLogsUrl` instead.
      operationId: getSnapshotBuildLogs
      parameters:
      - description: Snapshot ID
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: Whether to follow the logs stream
        explode: true
        in: query
        name: follow
        required: false
        schema:
          type: boolean
        style: form
      responses:
        "200":
          description: ""
      summary: Get snapshot build logs
      tags:
      - snapshots
  /snapshots/{id}/build-logs-url:
    get:
      operationId: getSnapshotBuildLogsUrl
      parameters:
      - description: Snapshot ID
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Url"
          description: The snapshot build logs URL
      summary: Get snapshot build logs URL
      tags:
      - snapshots
  /snapshots/{id}/activate:
    post:
      operationId: activateSnapshot
      parameters:
      - description: Snapshot ID
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SnapshotDto"
          description: The snapshot has been successfully activated.
      summary: Activate a snapshot
      tags:
      - snapshots
  /snapshots/{id}/deactivate:
    post:
      operationId: deactivateSnapshot
      parameters:
      - description: Snapshot ID
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: The snapshot has been successfully deactivated.
      summary: Deactivate a snapshot
      tags:
      - snapshots
  /preview/{sandboxId}/open:
    get:
      operationId: isSandboxPreviewOpen
      parameters:
      - description: ID of the sandbox
        explode: false
        in: path
        name: sandboxId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                type: boolean
          description: Open preview status of the sandbox
      summary: Check if sandbox preview is open
      tags:
      - preview
  /preview/{sandboxId}/validate/{authToken}:
    get:
      operationId: isValidAuthToken
      parameters:
      - description: ID of the sandbox
        explode: false
        in: path
        name: sandboxId
        required: true
        schema:
          type: string
        style: simple
      - description: Auth token of the sandbox
        explode: false
        in: path
        name: authToken
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                type: boolean
          description: Sandbox auth token validation status
      summary: Check if sandbox auth token is valid
      tags:
      - preview
  /preview/{sandboxId}/access:
    get:
      operationId: hasSandboxAccess
      parameters:
      - explode: false
        in: path
        name: sandboxId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                type: boolean
          description: User access status to the sandbox
      summary: Check if user has access to the sandbox
      tags:
      - preview
  /preview/{signedPreviewToken}/{port}/sandbox-id:
    get:
      operationId: getSandboxIdFromSignedPreviewUrlToken
      parameters:
      - description: Signed preview URL token
        explode: false
        in: path
        name: signedPreviewToken
        required: true
        schema:
          type: string
        style: simple
      - description: Port number to get sandbox ID from signed preview URL token
        explode: false
        in: path
        name: port
        required: true
        schema:
          type: number
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                type: string
          description: Sandbox ID from signed preview URL token
      summary: Get sandbox ID from signed preview URL token
      tags:
      - preview
...
components:
  schemas:
...
    SandboxClass:
      enum:
      - linux-vm
      - container
      - android
      - windows
      type: string
      x-enum-varnames:
      - LINUX_VM
      - CONTAINER
      - ANDROID
      - WINDOWS
...
    GpuType:
      description: The GPU type assigned to the sandbox
      enum:
      - H100
      - RTX-PRO-6000
      type: string
      x-enum-varnames:
      - H100
      - RTX_PRO_6000
...
    BuildInfo:
      properties:
        dockerfileContent:
          description: The Dockerfile content used for the build
          example: |-
            FROM node:14
            WORKDIR /app
            COPY . .
            RUN npm install
            CMD ["npm", "start"]
          type: string
        contextHashes:
          description: The context hashes used for the build
          example:
          - hash1
          - hash2
          items:
            type: string
          type: array
        createdAt:
          description: The creation timestamp
          format: date-time
          type: string
        updatedAt:
          description: The last update timestamp
          format: date-time
          type: string
        snapshotRef:
          description: The snapshot reference
          example: daytonaio/sandbox:latest
          type: string
      required:
      - createdAt
      - snapshotRef
      - updatedAt
      type: object
...
    CreateBuildInfo:
      properties:
        dockerfileContent:
          description: The Dockerfile content used for the build
          example: |-
            FROM node:14
            WORKDIR /app
            COPY . .
            RUN npm install
            CMD ["npm", "start"]
          type: string
        contextHashes:
          description: The context hashes used for the build
          example:
          - hash1
          - hash2
          items:
            type: string
          type: array
      required:
      - dockerfileContent
      type: object
...
    Url:
      example:
        url: url
      properties:
        url:
          description: URL response
          type: string
      required:
      - url
      type: object
...
    CreateSnapshot:
      example:
        disk: 3
        imageName: ubuntu:22.04
        memory: 1
        buildInfo: ""
        entrypoint: sleep infinity
        name: ubuntu-4vcpu-8ram-100gb
        cpu: 1
        gpu: 0
        gpuType:
        - H100
        sandboxClass: linux-vm
      properties:
        name:
          description: The name of the snapshot
          example: ubuntu-4vcpu-8ram-100gb
          type: string
        imageName:
          description: The image name of the snapshot
          example: ubuntu:22.04
          type: string
        entrypoint:
          description: The entrypoint command for the snapshot
          example: sleep infinity
          items:
            type: string
          type: array
        cpu:
          description: CPU cores allocated to the resulting sandbox
          example: 1
          type: integer
        gpu:
          description: GPU units allocated to the resulting sandbox
          example: 0
          type: integer
        gpuType:
          description: Preferred GPU type for the resulting sandbox.
          example:
          - H100
          items:
            $ref: "#/components/schemas/GpuType"
          type: array
        memory:
          description: Memory allocated to the resulting sandbox in GB
          example: 1
          type: integer
        disk:
          description: Disk space allocated to the sandbox in GB
          example: 3
          type: integer
        buildInfo:
          allOf:
          - $ref: "#/components/schemas/CreateBuildInfo"
          description: Build information for the snapshot
        sandboxClass:
          allOf:
          - $ref: "#/components/schemas/SandboxClass"
          description: Target sandbox class. Determines which runners can host sandboxes
            created from this snapshot.
          example: linux-vm
      required:
      - name
      type: object
    SnapshotState:
      enum:
      - building
      - pending
      - pulling
      - active
      - inactive
      - error
      - build_failed
      - removing
      type: string
    SnapshotDto:
      example:
        imageName: imageName
        buildInfo: ""
        cpu: 6.027456183070403
        gpu: 1.4658129805029452
        createdAt: 2000-01-23T04:56:07.000+00:00
        disk: 5.637376656633329
        ref: daytonaio/sandbox:latest
        size: 0.8008281904610115
        mem: 5.962133916683182
        lastUsedAt: 2000-01-23T04:56:07.000+00:00
        entrypoint:
        - entrypoint
        - entrypoint
        errorReason: errorReason
        name: name
        id: id
        state: ""
        initialRunnerId: runner123
        gpuType: H100
        updatedAt: 2000-01-23T04:56:07.000+00:00
        sandboxClass: linux-vm
      properties:
        id:
          type: string
        name:
          type: string
        imageName:
          type: string
        state:
          allOf:
          - $ref: "#/components/schemas/SnapshotState"
        size:
          nullable: true
          type: number
        entrypoint:
          items:
            type: string
          nullable: true
          type: array
        cpu:
          type: number
        gpu:
          type: number
        gpuType:
          allOf:
          - $ref: "#/components/schemas/GpuType"
          description: The GPU type assigned to the snapshot
          example: H100
        mem:
          type: number
        disk:
          type: number
        errorReason:
          nullable: true
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        lastUsedAt:
          format: date-time
          nullable: true
          type: string
        buildInfo:
          allOf:
          - $ref: "#/components/schemas/BuildInfo"
          description: Build information for the snapshot
        initialRunnerId:
          description: The initial runner ID of the snapshot
          example: runner123
          type: string
        ref:
          description: The snapshot reference
          example: daytonaio/sandbox:latest
          type: string
        sandboxClass:
          description: The sandbox class of the snapshot
          enum:
          - linux-vm
          - container
          - android
          - windows
          example: linux-vm
          type: string
          x-enum-varnames:
          - LINUX_VM
          - CONTAINER
          - ANDROID
          - WINDOWS
      required:
      - cpu
      - createdAt
      - disk
      - entrypoint
      - errorReason
      - gpu
      - id
      - lastUsedAt
      - mem
      - name
      - size
      - state
      - updatedAt
      type: object
    PaginatedSnapshots:
      example:
        total: 2.3021358869347655
        totalPages: 9.301444243932576
        page: 7.061401241503109
        items:
        - imageName: imageName
          buildInfo: ""
          cpu: 6.027456183070403
          gpu: 1.4658129805029452
          createdAt: 2000-01-23T04:56:07.000+00:00
          disk: 5.637376656633329
          ref: daytonaio/sandbox:latest
          size: 0.8008281904610115
          mem: 5.962133916683182
          lastUsedAt: 2000-01-23T04:56:07.000+00:00
          entrypoint:
          - entrypoint
          - entrypoint
          errorReason: errorReason
          name: name
          id: id
          state: ""
          initialRunnerId: runner123
          gpuType: H100
          updatedAt: 2000-01-23T04:56:07.000+00:00
          sandboxClass: linux-vm
        - imageName: imageName
          buildInfo: ""
          cpu: 6.027456183070403
          gpu: 1.4658129805029452
          createdAt: 2000-01-23T04:56:07.000+00:00
          disk: 5.637376656633329
          ref: daytonaio/sandbox:latest
          size: 0.8008281904610115
          mem: 5.962133916683182
          lastUsedAt: 2000-01-23T04:56:07.000+00:00
          entrypoint:
          - entrypoint
          - entrypoint
          errorReason: errorReason
          name: name
          id: id
          state: ""
          initialRunnerId: runner123
          gpuType: H100
          updatedAt: 2000-01-23T04:56:07.000+00:00
          sandboxClass: linux-vm
      properties:
        items:
          items:
            $ref: "#/components/schemas/SnapshotDto"
          type: array
        total:
          type: number
        page:
          type: number
        totalPages:
          type: number
      required:
      - items
      - page
      - total
      - totalPages
      type: object
...